feat: archive out-of-scope models to part2/archive/
Archive models not required by the spec:
  - Lasso & ElasticNet (spec requires Ridge OR Lasso, keeping Ridge)
  - Bayesian Linear Regression (bonus/optional, keeping Kernel Ridge instead)
  - Decision Tree & Random Forest (not mentioned in spec)

Archive structure:
  part2/archive/
    lasso_elasticnet_models.py   — sklearn Lasso/ElasticNet with StandardScaler
    bayesian_random_forest_models.py — from-scratch Bayesian, DecisionTree, RF
    test_archived_models.py      — all 14 archived tests (passing)

Source files stripped:
  model_comparison.py    — OLS + Ridge only (no Lasso/ElasticNet)
  advanced_methods.py    — Kernel Ridge only (no Bayesian/RF/DT)
  test files updated accordingly
